Moved profile change functions to appropriate fragment.
